1. Added model file interface for DEPTH_CONVERSION 2. Small fixes here and there
git-svn-id: https://code.nr.no/scm/svn/crava/trunk@119 176de4d3-4747-0410-aa96-47f6d6634214

+++ nrlib/segy/segy.cpp
@@ -132,7 +132,7 @@ SegY::readAllTraces(Volume *volume, double zPad, bool onlyVolume)
   LogKit::LogMessage(LogKit::LOW,"\n  ^");
   int traceSize = datasize_*nz_+240;
   long long bytesRead = 3600+traceSize;
-  for(unsigned int i=1 ; i<nTraces_ ; i++)
+  for(unsigned int i=1 ; i < static_cast<unsigned int>(nTraces_) ; i++)
   {
     double percentDone = bytesRead/static_cast<double>(fSize);
     if(percentDone > nextWrite)
